verbify-ai-backend/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
soundfile
google-generativeai
requests

